cepheus: power: rename to avoid conflict with hardware/xiaomi
Lineage is also moving to Pixel's libperfmgr, and they also named
it as android.hardware.power-service.xiaomi-libperfmgr which will
unfortunately cause a name duplication as we also track lineage's
hardware/xiaomi.

Rename ours to android.hardware.power-service.cepheus-libperfmgr
so our in-tree HAL can co-exist with the one in hardware/xiaomi.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>